Modularized request forwarding functionality of server to improve ability to test code
